requirements.txt
kafka-python==2.0.2
zstandard==0.21.0
lz4==4.3.2
Flask==2.2.2
dacite==1.7.0
redis==4.5.4
flasgger==0.9.5
Flask-Cors==3.0.10
flask-swagger-ui==4.11.1
urllib3==1.26.4
PyYAML==6.0